test: setup shared coverage report test utility and test switch/case
